JobRepositoryImpl.kt remade on suspend + Retrofit

+++ app/src/main/java/ru/netology/papillon/repository/JobRepositoryImpl.kt
@@ -1,24 +1,63 @@
 package ru.netology.papillon.repository
 
-import androidx.lifecycle.Transformations
+import androidx.lifecycle.map
+import ru.netology.papillon.api.Api
 import ru.netology.papillon.dao.JobDao
 import ru.netology.papillon.dto.Job
 import ru.netology.papillon.entity.JobEntity
+import ru.netology.papillon.error.ApiError
+import ru.netology.papillon.error.NetworkError
+import ru.netology.papillon.error.UnknownError
+import ru.netology.papillon.extensions.toDtoJob
+import ru.netology.papillon.extensions.toEntityJob
+import java.io.IOException
 
-class JobRepositoryImpl(
-    private val jobDao: JobDao
-) : JobRepository {
-    override fun getAllJobs() = Transformations.map(jobDao.getAll()) {
-        it.map {
-            Job(it.id, it.company, it.position, it.start, it.finish, it.link)
+class JobRepositoryImpl(private val jobDao: JobDao): Repository<Job> {
+
+    override val data = jobDao.getAll().map(List<JobEntity>::toDtoJob)
+
+    override suspend fun getAll() {
+        try {
+            val response = Api.service.getAllJobs()
+            if (!response.isSuccessful) {
+                throw ApiError(response.code(), response.message())
+            }
+            val body = response.body() ?: throw ApiError(response.code(), response.message())
+            jobDao.insert(body.toEntityJob())
+        } catch (e: IOException) {
+            NetworkError
+        } catch (e: Exception) {
+            UnknownError
         }
     }
 
-    override fun removedById(id: Long) {
-        jobDao.removedById(id)
+    override suspend fun save(t: Job) {
+        try {
+            val response = Api.service.saveJob(t)
+            if (!response.isSuccessful) {
+                throw ApiError(response.code(), response.message())
+            }
+            val body = response.body() ?: throw ApiError(response.code(), response.message())
+            jobDao.insert(JobEntity.fromDtoJob(body))
+        } catch (e: IOException) {
+            NetworkError
+        } catch (e: Exception) {
+            UnknownError
+        }
     }
 
-    override fun saveJob(job: Job) {
-        jobDao.saveJob(JobEntity.fromDtoJob(job))
+    override suspend fun removedById(id: Long) {
+        try {
+            val response = Api.service.removedByIdJob(id)
+            if (!response.isSuccessful) {
+                throw ApiError(response.code(), response.message())
+            }
+            response.body() ?: throw ApiError(response.code(), response.message())
+            jobDao.removedById(id)
+        } catch (e: IOException) {
+            NetworkError
+        } catch (e: Exception) {
+            UnknownError
+        }
     }
 }
